requirements.txt
langchain
langchain-community
langchain_groq 
# create vector store
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain_huggingface


